Remove `eq_constraints` (#1064)
* remove file

* full excision

* update builder
