+++ bluemira/equilibria/opt_problems.py
@@ -42,10 +42,9 @@ import numpy as np
 
 import bluemira.equilibria.opt_objectives as objectives
 from bluemira.equilibria.coils import CoilSet
-from bluemira.equilibria.eq_constraints import MagneticConstraintSet
 from bluemira.equilibria.equilibrium import Equilibrium
 from bluemira.equilibria.error import EquilibriaError
-from bluemira.equilibria.opt_constraints import MagneticConstraint
+from bluemira.equilibria.opt_constraints import MagneticConstraint, MagneticConstraintSet
 from bluemira.equilibria.positioner import RegionMapper
 from bluemira.utilities.opt_problems import (
     OptimisationConstraint,
