+++ documentation/source/equilibria/doc_parametric_constraints.py
@@ -1,4 +1,4 @@
-from bluemira.equilibria.eq_constraints import (
+from bluemira.equilibria.opt_constraints import (
     FieldNullConstraint,
     IsofluxConstraint,
     MagneticConstraintSet,
